# service-api-gateway-node/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FairGiG Grievance Node API
  description: |
    Express service for grievances, analytics, and certificate endpoints.
    Interactive docs are available at /api-docs.
    Grievances, analytics, certificates, and worker community require `Authorization: Bearer` (Supabase access token).
  version: 0.2.0
servers:
  - url: http://localhost:5000
    description: Local
  - url: /
    description: Same-origin host
...
paths:
...
  /api/grievances:
    get:
      tags: [Grievances]
      summary: List grievances
      operationId: listGrievances
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: worker_id
          description: Verifier/advocate/analyst only; ignored for workers (always scoped to caller).
          schema:
            type: string
        - in: query
          name: platform
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [open, escalated, resolved]
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile
        "200":
          description: Grievances list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievanceList"
    post:
      tags: [Grievances]
      summary: Create grievance
      operationId: createGrievance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, category, description]
              properties:
                worker_id:
                  type: string
                  description: Required for verifier/advocate/analyst; ignored for workers (uses JWT subject).
                platform:
                  type: string
                category:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [open, escalated, resolved]
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile or worker cannot set worker_id for others
        "201":
          description: Grievance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievance"
  /api/grievances/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    get:
      tags: [Grievances]
      summary: Get grievance by id
      operationId: getGrievanceById
      security:
        - bearerAuth: []
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile
        "404":
          description: Not found or not visible to this role
        "200":
          description: Grievance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievance"
    patch:
      tags: [Grievances]
      summary: Update grievance
      operationId: updateGrievance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  type: string
                category:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [open, escalated, resolved]
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile
        "404":
          description: Not found or not owned
        "200":
          description: Updated grievance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievance"
    delete:
      tags: [Grievances]
      summary: Delete grievance
      operationId: deleteGrievance
      security:
        - bearerAuth: []
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile
        "404":
          description: Not found or not owned
        "200":
          description: Delete confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeDeleted"
  /api/grievances/{id}/tags:
    post:
      tags: [Grievances]
      summary: Add tags to grievance
      operationId: addGrievanceTags
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tags]
              properties:
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile
        "404":
          description: Not found or not owned
        "200":
          description: Updated grievance with merged tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievance"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase user access_token (not service_role)
  schemas:
    Grievance:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        platform:
          type: string
        category:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, escalated, resolved]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    EnvelopeGrievance:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            grievance:
              $ref: "#/components/schemas/Grievance"
        error:
          nullable: true
    EnvelopeGrievanceList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Grievance"
            pagination:
              type: object
              properties:
                limit:
                  type: integer
                offset:
                  type: integer
                count:
                  type: integer
        error:
          nullable: true
    EnvelopeDeleted:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            deleted:
              type: boolean
            id:
              type: string
        error:
          nullable: true
...